Avoid the overwriting of a dotenv file

+++ demostart.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-cp .env_sample .env
+if [ ! -f ".env" ]; then
+  cp .env_sample .env
+fi
 docker-compose up -d
